Use existing `SettingRow` to render preference rows
